+++ src/einspline/bspline_structs.h
@@ -7,6 +7,7 @@
 #ifndef BSPLINE_STRUCTS_STD_H
 #define BSPLINE_STRUCTS_STD_H
 #include <stdlib.h>
+#include "bspline_base.h"
 
 ///////////////////////////
 // Single precision real //
